Allow control of AcoustID API meta parameter
Added `options.meta` that will be passed on to the AcoustID Web Service
API as the *meta* parameter. If this option is not given, it defaults to
returning all meta data (see `META_DEFAULT` variable).

Added test that checks for expected behavior of using a meta parameter.

Documented usage of `option.meta` in README.md.